Patch Set #1
This set contains miscellaneous fixes for memory leaks, race conditions
and corner cases that could have caused the RMA implementation to crash or not
to release all the acquired memory.
It brings the source code in sync with 13b0cbaac15c7a305a907374ed3bab1c900a2589
of the main private repository.